moved IDLE logic to Sequence and Fallback
The fact that an action can be ticked only if RUNNING or IDLE was
confusing.
But it was a way to deal with an issue in SequenceNode and FallbackNode.

But since the latter are the only special case, the same logic was moved
there. Brakes incapsulation, but those nodes deserve to die anyway...