Modify documentation images.
SequenceNode:
AimTo -> AimAt
"Aim at a target" might sound more appropriate in this example than "aim to reach a goal".

SequenceStar:
Sequence -> ReactiveSequence
The text says "On the other hand, isBatteryOK must be checked at every tick,
for this reason its parent must be a ReactiveSequence."
Modify the image to match the text.