docs: clearer code for first tutorial
Now at the end of the tutorial the entire code that can be copied and pasted
directy to test the demo is provided.
In particular the needed include files are added which are not mentioned along
the tutorial and this spare a little time for new users to figure them out.