docs: Small changes to tutorial 02 (#225)
* docs: fix to my_tree.xml

Before my_tree.xml was containing an `SayHello` Action that was not registered
inside the BehaviourTreeFactory contained in the simple_bt.cpp

* docs: tutorial 1 my_tree.xml editable in Groot

Add to the xml tree definition the node declaration needed to edit the behavior
xml file with groot

* docs: clearer code for first tutorial

Now at the end of the tutorial the entire code that can be copied and pasted
directy to test the demo is provided.
In particular the needed include files are added which are not mentioned along
the tutorial and this spare a little time for new users to figure them out.

* docs: remove blank lines

* docs: little addition to tutorial_02 docs

- Provide a default implementation of `SaySomething2`
- Add Groot support in the xml file

* docs: Add missing SaySimple2

* docs: removed not useful stuff

Co-authored-by: Valerio Magnago <valerio.magnago@fraunhofer.it>