Fixes for compilation on windows. (#248)
* Fix for detecting ZeroMQ on windows

Naming convention is a bit different for ZeroMQ, specifically on Windows with vcpkg. While ZMQ and ZeroMQ are valid on linux, the ZMQ naming convention only works on linux.

* Compilation on windows not working with /WX

* Macro collision on Windows

On windows, the macros defined in the abstract logger collides with other in windows.h. Made them lowercase to avoid collision